V1
Implemented calls to the LLM using a specific strucure.

Implemented chat roles and chat history.

Implemented an extractPokemonName function, that retrieves the Pokémon's name from the initial response.

Build a simple front-end.

Implemented a fetch call to the PokéAPI that used the extracted Pokémon name. This fetches an image of the requested Pokémon.

Implemented two input forms. One initial form that asks for a Pokémon name. This gets used in a engineered prompt to generate a competitive moveset. This first form then gets replaced with a second form. The user can use this to interact with the LLM and request changes for the generated moveset.

Implemented functions that check and trims the given response, depending on if it is a moveset or not. There's two div's in my front-end for responses. Movesets get placed in one div, any other information get placed in the other.

Implemented some basic styling.